New workflow using LuaRooms
Each lua domain now runs inside a LuaRoom. With this new way
we will be able to manage scenes and lua behaviours easily.

The basic idea is that each scene will usually have one LuaRoom that will
set up the scene and the agents and the manage them.

This way this new UDMS can have folders (like the original one) that will
completely set up and run scenes.